Add special case for Apple developer portal login.
Note that apple does not allow submit of the login form without user
actual key interaction. This means that after the field username/pass
are filled the user still needs to focus each field and type a letter
before the submit button is enabled.

I have been unable to automatically trigger the event that enables the
submit button programmatically.